Display keybinding in the tooltip
A little known feature of Atom tooltip is that it supports displaying the keybinding directly inside of the tooltip if you pass a `keyBindingCommand` value as the atom command.

In Nuclide we're already giving the command so I figured it would be cool to display the keybinding as well :)

Let me know if it's something you'd consider! Thanks