Combine string and function callback for button view
I changed the "executeCallback" function so that the user can use both
strings (for atom command dispatching) and function (for whatever the
user wants) when the `callback` option is set to an array.

To achieve this I convert the `callback` value to an array if it is not
and then iterate over the array and perform either the dispatch or
function call as it happened before.