Refactor raf-debounce, reduce length and complexity
Update the implementation of raf-debounce to reduce it's length and 
reduce complecity. This is achieve primarily through the removal of 
if-statements that are "unnecessary" in the sense that everything still 
works when they're removed.

Most notably, there is no longer a check that the requestID is set, as 
it is safe to call `window.cancelAnimationFrame` with `null` or 
`undefined`.