Refactor updatePosition, reduce length and complexity
Reduce the complexity of the `updatePosition` method of the ToolBarView 
class primarily by extracting the logic that creates a new panel into a 
separate method.

This already reduced the length of the `updatePosition` method. The 
length is further reduced by rewriting the logic that sets the classes 
for the new panel element.