Add simple slide-in animation for the tool-bar
I used @keyframes to create the animations. The animations are based on 
margins with exact pixel values given the configured tool-bar icon size.

If the animations are created starting from, e.g. `margin-left: -100%` 
the toolbar is hidden, but the panel is not and so the animation does 
not work.

If the animations are created using the CSS `transform` property (I 
tried both `scale()` and `translate()`) you have the same problem.

I'm not sure how this animation works if the toolbar somehow has 
multiple rows/columns...